Refactor to angular 11 library
* transformed to angular 11 lib
* removed unnecessary files
* fixed libs dependencies
* fixed html-to-image
* fixed feature labels
* fixed dimensional issues
* code leaved as is - future purification needed
