--- .npmignore
@@ -1,4 +1,0 @@
-src
-tsconfig.json
-tslint.json
-.prettierrc
