--- .nvmrc
@@ -1,1 +1,0 @@
-lts/dubnium
