--- CONTRIBUTING.md
@@ -1,40 +1,0 @@
-# Contributing
-
-## Getting started
-
-Before you begin:
-- Check out the [existing issues](https://github.com/BioComputingUP/feature-viewer-typescript/issues).
-
-### Don't see your issue? Open one
-
-If you spot something new, open an issue. We'll use the issue to have a conversation about the problem you want to fix.
-
-### Ready to make a change? Fork the repo
-
-Fork using GitHub Desktop:
-
-- [Getting started with GitHub Desktop](https://docs.github.com/en/desktop/installing-and-configuring-github-desktop/getting-started-with-github-desktop) will guide you through setting up Desktop.
-- Once Desktop is set up, you can use it to [fork the repo](https://docs.github.com/en/desktop/contributing-and-collaborating-using-github-desktop/cloning-and-forking-repositories-from-github-desktop)!
-
-Fork using the command line:
-
-- [Fork the repo](https://docs.github.com/en/github/getting-started-with-github/fork-a-repo#fork-an-example-repository) so that you can make your changes without affecting the original project until you're ready to merge them.
-
-Fork with [GitHub Codespaces](https://github.com/features/codespaces):
-
-- [Fork, edit, and preview](https://docs.github.com/en/free-pro-team@latest/github/developing-online-with-codespaces/creating-a-codespace) using [GitHub Codespaces](https://github.com/features/codespaces) without having to install and run the project locally.
-
-### Make your update:
-Make your changes to the file(s) you'd like to update. Here are some tips and tricks for [using the docs codebase](#working-in-the-githubdocs-repository).
-  - Are you making changes to the application code? You'll need **Node.js lts/dubnium** to run the library locally. 
-
-### Open a pull request
-When you're done making changes and you'd like to propose them for review, open your pull request (PR).
-
-### Submit your PR & get it reviewed
-- Once you submit your PR, we will review it with you. The first thing you're going to want to do is a [self review](#self-review).
-- After that, we may have questions, check back on your PR to keep up with the conversation.
-
-### Your PR is merged!
-Congratulations! The whole GitHub community thanks you. :sparkles:
-Once your PR is merged, you will be proudly listed as a contributor in the [contributor chart](https://github.com/BioComputingUP/feature-viewer-typescript/graphs/contributors).
