+++ angular.json
@@ -1,0 +1,45 @@
+{
+  "$schema": "./node_modules/@angular/cli/lib/config/schema.json",
+  "version": 1,
+  "newProjectRoot": "projects",
+  "projects": {
+    "ngx-feature-viewer": {
+      "projectType": "library",
+      "root": "projects/ngx-feature-viewer",
+      "sourceRoot": "projects/ngx-feature-viewer/src",
+      "prefix": "lib",
+      "architect": {
+        "build": {
+          "builder": "@angular-devkit/build-angular:ng-packagr",
+          "options": {
+            "tsConfig": "projects/ngx-feature-viewer/tsconfig.lib.json",
+            "project": "projects/ngx-feature-viewer/ng-package.json"
+          },
+          "configurations": {
+            "production": {
+              "tsConfig": "projects/ngx-feature-viewer/tsconfig.lib.prod.json"
+            }
+          }
+        },
+        "test": {
+          "builder": "@angular-devkit/build-angular:karma",
+          "options": {
+            "main": "projects/ngx-feature-viewer/src/test.ts",
+            "tsConfig": "projects/ngx-feature-viewer/tsconfig.spec.json",
+            "karmaConfig": "projects/ngx-feature-viewer/karma.conf.js"
+          }
+        },
+        "lint": {
+          "builder": "@angular-eslint/builder:lint",
+          "options": {
+            "lintFilePatterns": [
+              "projects/ngx-feature-viewer/**/*.ts",
+              "projects/ngx-feature-viewer/**/*.html"
+            ]
+          }
+        }
+      }
+    }
+  },
+  "defaultProject": "ngx-feature-viewer"
+}
