--- assets/fv.scss.d.ts
@@ -1,17 +1,0 @@
-declare const styles: {
-  readonly "helperButton": string;
-  readonly "selectionRect": string;
-  readonly "brush": string;
-  readonly "selection": string;
-  readonly "mybutton": string;
-  readonly "mybuttonsquare": string;
-  readonly "mybuttoncircle": string;
-  readonly "axis": string;
-  readonly "arrow": string;
-  readonly "fvtooltip": string;
-  readonly "fvcustomtooltip": string;
-  readonly "pageoverlay": string;
-  readonly "loadingcolorflag": string;
-};
-export = styles;
-
