+++ package.json
@@ -1,56 +1,59 @@
 {
-  "name": "feature-viewer-webpack",
-  "version": "2.3.0",
-  "description": "",
-  "main": "index.js",
-  "author": "Lisanna Paladin",
+  "name": "ngx-feature-viewer",
+  "version": "0.0.0",
+  "description": "Represent biological data with the feature viewer library.",
   "scripts": {
-    "sass": "sass assets/fv.scss dist/fv.css",
-    "tsc": "tsc --watch",
-    "build": "npm run sass && webpack"
+    "start": "ng serve",
+    "build": "ng build",
+    "build:prod": "ng build --prod",
+    "lint": "eslint './projects/ngx-feature-viewer/src/**'",
+    "lint:fix": "eslint --fix './projects/ngx-feature-viewer/src/**'"
   },
-  "files": [
-    "lib/**/*",
-    "DOC",
-    "assets",
-    "fonts"
-  ],
+  "private": true,
   "dependencies": {
-    "@types/d3": "5.7.2",
-    "@types/underscore": "1.8.16",
-    "byots": "4.1.0-dev.20200804.14.47",
-    "codemirror": "5.46.0",
-    "css-to-ts": "^1.0.1",
-    "d3-axis": "^1.0.12",
-    "d3-brush": "^1.0.6",
-    "d3-scale": "2.2.2",
-    "d3-selection": "1.4.0",
-    "d3-shape": "1.3.5",
-    "html-to-image": "^0.1.1",
-    "npx": "^10.2.2",
-    "sass": "^1.32.12",
-    "typed-css-modules": "^0.6.4",
-    "underscore": "^1.9.1"
+    "@angular/animations": "^11.1.2",
+    "@angular/common": "^11.1.2",
+    "@angular/compiler": "^11.1.2",
+    "@angular/core": "^11.1.2",
+    "@angular/forms": "^11.1.2",
+    "@angular/platform-browser": "^11.1.2",
+    "@angular/platform-browser-dynamic": "^11.1.2",
+    "@angular/router": "^11.1.2",
+    "eslint-config-google": "^0.14.0",
+    "rxjs": "^6.6.0",
+    "tslib": "^2.0.0",
+    "zone.js": "^0.11.3"
   },
-  "keywords": [],
-  "license": "ISC",
   "devDependencies": {
-    "@types/d3-axis": "1.0.12",
-    "@types/d3-scale": "2.1.1",
-    "@types/d3-selection": "1.4.1",
-    "@types/d3-shape": "1.3.1",
-    "aws-sdk": "2.455.0",
-    "css-loader": "^5.0.0",
-    "jsdoc": "^3.6.3",
-    "material-design-icons": "^3.0.1",
-    "material-design-lite": "^1.3.0",
-    "node-sass": "4.14.1",
-    "sass-loader": "^7.1.0",
-    "ts-loader": "^9.1.1",
-    "tsc-watch": "^4.2.9",
-    "typescript": "^4.2.4",
-    "typings-for-css-modules-loader": "^1.7.0",
-    "webpack": "^5.36.2",
-    "webpack-cli": "^4.6.0"
+    "@angular-devkit/build-angular": "^0.1101.4",
+    "@angular-eslint/builder": "1.2.0",
+    "@angular-eslint/eslint-plugin": "1.2.0",
+    "@angular-eslint/eslint-plugin-template": "1.2.0",
+    "@angular-eslint/schematics": "1.2.0",
+    "@angular-eslint/template-parser": "1.2.0",
+    "@angular/cli": "^11.1.4",
+    "@angular/compiler-cli": "^11.1.2",
+    "@types/jasmine": "^3.6.0",
+    "@types/node": "^12.11.1",
+    "@typescript-eslint/eslint-plugin": "4.3.0",
+    "@typescript-eslint/parser": "4.3.0",
+    "codelyzer": "^6.0.0",
+    "eslint": "^7.6.0",
+    "eslint-plugin-html": "^6.2.0",
+    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-jsdoc": "30.7.6",
+    "eslint-plugin-prefer-arrow": "1.2.2",
+    "jasmine-core": "^3.6.0",
+    "jasmine-spec-reporter": "^5.0.0",
+    "karma": "^5.2.0",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-coverage": "^2.0.3",
+    "karma-jasmine": "^4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "ng-packagr": "^11.0.0",
+    "protractor": "^7.0.0",
+    "ts-node": "^8.3.0",
+    "tslint": "^6.1.0",
+    "typescript": "^4.1.2"
   }
 }
