+++ projects/ngx-feature-viewer/package.json
@@ -1,0 +1,47 @@
+{
+  "name": "ngx-feature-viewer",
+  "version": "0.0.1",
+  "description": "Represent biological data with the feature viewer library",
+  "peerDependencies": {
+    "@angular/common": "^11.1.2",
+    "@angular/core": "^11.1.2"
+  },
+  "dependencies": {
+    "@types/d3": "5.7.2",
+    "@types/underscore": "1.8.16",
+    "byots": "4.1.0-dev.20200804.14.47",
+    "codemirror": "5.46.0",
+    "css-to-ts": "^1.0.1",
+    "d3-axis": "^1.0.12",
+    "d3-brush": "^1.0.6",
+    "d3-scale": "2.2.2",
+    "d3-selection": "1.4.0",
+    "d3-shape": "1.3.5",
+    "html-to-image": "1.9.0",
+    "npx": "^10.2.2",
+    "sass": "^1.32.12",
+    "typed-css-modules": "^0.6.4",
+    "underscore": "^1.9.1"
+  },
+  "keywords": [],
+  "license": "GPL-2.0",
+  "devDependencies": {
+    "@types/d3-axis": "1.0.12",
+    "@types/d3-scale": "2.1.1",
+    "@types/d3-selection": "1.4.1",
+    "@types/d3-shape": "1.3.1",
+    "aws-sdk": "2.455.0",
+    "css-loader": "^5.0.0",
+    "jsdoc": "^3.6.3",
+    "material-design-icons": "^3.0.1",
+    "material-design-lite": "^1.3.0",
+    "node-sass": "4.14.1",
+    "sass-loader": "^7.1.0",
+    "ts-loader": "^9.1.1",
+    "tsc-watch": "^4.2.9",
+    "typescript": "^4.2.4",
+    "typings-for-css-modules-loader": "^1.7.0",
+    "webpack": "^5.36.2",
+    "webpack-cli": "^4.6.0"
+  }
+}
