+++ projects/ngx-feature-viewer/src/lib/custom-d3.ts
@@ -1,0 +1,5 @@
+export * as D3Selection from 'd3-selection';
+export * as D3Scale from 'd3-scale';
+export * as D3Shape from 'd3-shape';
+export * as D3Axis from 'd3-axis';
+export * as D3Brush from 'd3-brush';
