+++ projects/ngx-feature-viewer/src/lib/index.ts
@@ -1,0 +1,1 @@
+export * from './feature-viewer';
