+++ projects/ngx-feature-viewer/src/public-api.ts
@@ -1,0 +1,5 @@
+/*
+ * Public API Surface of ngx-feature-viewer
+ */
+
+export * from './lib/feature-viewer';
