--- src/custom-d3.ts
@@ -1,5 +1,0 @@
-export * from 'd3-selection';
-export * from 'd3-scale';
-export * from 'd3-shape';
-export * from 'd3-axis';
-export * from 'd3-brush';
