--- test-index.html
@@ -1,24 +1,0 @@
-<head>
-<link rel="stylesheet" type="text/css" href="./dist/fv.css"></div>
-</head>
-
-Feature Viewer in plain html
-<div id="fvDivInit"></div>
-
-<script src="./dist/feature-viewer-webpack.bundle.js"></script>
-<script>
-    const initViewer = function (divId) {
-        // tslint:disable-next-line:max-line-length
-        var P04050 = 'MTKFTILLISLLFCIAHTCSASKWQHQQDSCRKQLQGVNLTPCEKHIMEKIQGRGDDDDDDDDDNHILRTMRGRINYIRRNEGKDEDEEEEGHMQKCCTEMSELRSPLMTKFTILLISLLFCIAHTCSASKWQHQQDSCRKQLQGVNLTPCEKHIMEKIQGRGDDDDDDDDDNHILRTMRGRINYIRRNEGKDEDEEEEGHMQKCCTEMSELRSPLMTKFTILLISLLFCIAHTCSASKWQHQQDSCRKQLQGVNLTPCEKHIMEKIQGRGDDDDDDDDDNHILRTMRGRINYIRRNEGKDEDEEEEGHMQKCCTEMSELRSPL';
-        var fv = new FeatureViewer(P04050, '#' + divId, {
-            toolbar: true,
-            toolbarPosition: 'left',
-            brushActive: true,
-            zoomMax: 10,
-            flagColor: '#DFD5F5'
-        });
-        return fv;
-    }
-    initViewer("fvDivInit")
-</script>
-
