--- webpack.config.js
@@ -1,32 +1,0 @@
-const path = require( "path" );
-
-module.exports = {
-
-    // bundling mode
-    mode: "production",
-
-    // entry files
-    entry: "./src/index.ts",
-
-    // output bundles (location)
-    output: {
-        path: path.resolve( __dirname, "dist" ),
-        filename: "feature-viewer-webpack.bundle.js",
-    },
-
-    // file resolutions
-    resolve: {
-        extensions: [ ".ts", ".js" ],
-    },
-
-    // loaders
-    module: {
-        rules: [
-            {
-                test: /\.tsx?/,
-                use: "ts-loader",
-                exclude: /node_modules/,
-            }
-        ]
-    }
-};
